+++ lib/hycop/enum/model_enums.dart
@@ -10,6 +10,7 @@ enum ExModelType {
   contents,
   user,
   team,
+  channel,
   watchHistory,
   playlist,
   end;